[Refact] refactor folder structure
