--- src/components/atoms/NavButton/index.module.css
@@ -1,15 +1,0 @@
-.button {
-  padding: 45px 42px;
-  margin: 0 20px 0 0;
-
-  font-weight: bold;
-  font-size: 20px;
-  text-align: center;
-
-  color: #000000;
-  background: #ffffff;
-
-  word-break: keep-all;
-  white-space: nowrap;
-  cursor: pointer;
-}
