--- src/components/molecules/Header/index.module.css
@@ -1,21 +1,0 @@
-.header__wrapper {
-  box-shadow: 0px 4px 10px rgba(0, 0, 0, 0.1);
-}
-
-.header__container {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-
-  max-width: 1080px;
-  margin: 0 auto;
-}
-
-.header__container--left {
-  display: flex;
-  align-items: center;
-}
-
-.header__container--right {
-  display: flex;
-}
