--- src/components/molecules/SearchBar/index.module.css
@@ -1,22 +1,0 @@
-.searchbar__wrapper {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-
-  max-width: 420px;
-  min-width: 200px;
-  height: 100%;
-  padding: 20px;
-
-  background-color: #eeeeee;
-  border-radius: 100px;
-}
-
-.searchbar__input {
-  font-style: normal;
-  font-weight: normal;
-  font-size: 20px;
-  line-height: 30px;
-
-  color: #7a7a7a;
-}
