--- src/components/templates/Home/index.module.css
@@ -1,11 +1,0 @@
-.home__wrapper {
-  min-height: 100vh;
-}
-
-.home__container {
-  max-width: 1080px;
-  margin: 0 auto;
-
-  display: flex;
-  flex-direction: column;
-}
