+++ src/index.css
@@ -1,5 +1,3 @@
-@import url(https://spoqa.github.io/spoqa-han-sans/css/SpoqaHanSansNeo.css);
-
 body {
   margin: 0;
   font-family: 'Spoqa Han Sans', -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
