--- src/pages/MainPage.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-function MainPage() {
-  return <div>MainPage</div>;
-}
-
-export default MainPage;
